refactored some stuff
